Feat: Table restyle
