+++ src/app/App.tsx
@@ -4,7 +4,7 @@ import { MainMap } from '@/app/features/MainMap';
 import store, { AppDispatch, RootState } from '@/lib/state/store';
 import { MapProvider } from '@/app/contexts/MapContexts';
 import { MAP_ID as MAIN_MAP_ID } from '@/app/features/MainMap/config';
-import { SidePanel } from '@/app/features/SidePanel';
+import SidePanel from '@/app/features/SidePanel';
 import Table from '@/app/features/Table';
 import { MapTools } from '@/app/features/MapTools';
 import { setShowSidePanel } from '@/lib/state/main/slice';
@@ -33,7 +33,14 @@ export const App: React.FC<Props> = (props) => {
         <>
             <HelpModal />
             <div className="flex">
-                <div id="side-panel-control" className={`fixed top-2 left-2`}>
+                <div
+                    id="side-panel-control"
+                    className={`fixed left-2 ${
+                        view === 'table'
+                            ? 'top-[unset] bottom-6 lg:top-2 lg:bottom-[unset]'
+                            : ''
+                    }`}
+                >
                     {!showSidePanel && (
                         <IconButton
                             onClick={() => handleSidePanelControlClick()}
@@ -45,11 +52,7 @@ export const App: React.FC<Props> = (props) => {
                 <div
                     id="side-panel"
                     className={`
-                      w-[70vw] lg:w-[45vw] xl:w-[30vw] 2xl:w-[20vw] 
-                     min-w-[300px] sm:min-w-[400px]
-                     max-w-[300px] sm:max-w-[400px]
-                     h-fit lg:h-full
-                     max-h-[88vh] lg:max-h-none
+                     w-[400px] max-w-[90svw]
                      flex overflow-hidden bg-primary
                      border lg:border-l-0 lg:border-t-0 lg:border-b-0
                      m-2 lg:m-0
