+++ src/app/components/common/Collapsible.tsx
@@ -24,11 +24,12 @@ const Collapsible: React.FC<Props> = (props) => {
         <div className="bg-primary text-black ">
             <button
                 title={`${isOpen ? 'Hide' : 'Show'} ${title}`}
-                className={`sticky top-0 w-full 
+                className={`sticky top-0 w-full -mt-1
                         flex justify-between items-center p-4 bg-primary-opaque 
                         font-bold text-lg 
                         border-t ${isOpen && 'border-b'} border-gray-300
-                        hover:bg-primary-opaque-hover focus:bg-primary-opaque-hover`}
+                        hover:bg-primary-opaque-hover focus:bg-primary-opaque-hover
+                        z-[1]`}
                 onClick={toggleCollapse}
             >
                 {title}
