+++ src/app/globals.css
@@ -48,12 +48,15 @@ div#tools {
 }
 
 div#scrollable-side-panel {
-    height: calc(100svh - 339px);
-    max-height: calc(100svh - 339px);
+    max-height: calc(100svh - 250px);
 }
 
 div#table-wrapper {
-    max-height: calc(100svh - 90px);
+    max-height: calc(100svh - 98px);
+}
+
+thead#table-header {
+    box-shadow: inset 0 0px 0 #00000000, inset 0 -2px 0 #d1d5db;
 }
 
 #date-range-columns {
@@ -109,8 +112,10 @@ a:active {
     div#scrollable-side-panel {
         height: calc(100svh - 219px);
         max-height: calc(100svh - 219px);
-        /* height: calc(100svh - 98px);
-        max-height: calc(100svh - 98px); */
+    }
+
+    div#table-wrapper {
+        max-height: calc(100svh - 75px);
     }
 }
